The python-packaging branch removes the global SETTINGS dictionary that was
being populated with some hard coded defaults in pudl.settings. The current
arrangement isn't finished, but the paths are now changeable, either by
setting paths in a user's `.pudl.yml` configuration file, or by passing in
`pudl_in` and `pudl_out` paths for the top level directories of the PUDL
inputs and outputs in the `pudl.settings.init()` function. This will also
allow us to get rid of the 'testing" databases and associated flags, since we
can just output the test databases or datapackages to a temporary testing area.
See #342.

The move away from global dependencies should also make it easier to tell what
a function is really relying on in order to do its job. Hopefully it will
make the code more readable and debugging easier, and allow easier re-use of
functions that were previously tied to hard-coded values.

Now instead of the global SETTINGS dictionary, `pudl.settings.init()` returns
a settings dictionary, depending only on the the `pudl_in` and `pudl_out`
paths that the user specifies.

The paths in the `pudl_settings` dictionary can be used to run
`pudl.settings.setup()` which creates a directory structure that looks pretty
similar to what we’ve got right now, and copies some default files over from
the Secret Place (see below). If `pudl.settings.setup()` gets no inputs, it’ll
run `pudl.settings.init()` itself, and try to pull the input and output root
directories from the `.pudl.yml` user settings file.

That directory structure looks like...

### `pudl_in:`
 * Contains the datastore, which is populated by our scripts, and organized by
   data source (`pudl_in/data/agency/form`)
 * Also contains a settings directory, which includes example/default settings
   files for use with the scripts -- however these are meant to be renamed and
   edited by the user. They’re just examples. No input file is ever referred to
   directly or by default in the code, except for the `.pudl.yml` that
   determines where these data and settings are deployed.

### `pudl_out`:
 * Contains directories organize by type of output:
   * `datapackage`
   * `parquet`
   * `sqlite`
 * Also contains directories for useful / necessary things distributed with the
   package that the user might need to edit or play currently including:
   * `notebooks` (including our tutorial notebooks)

### `scripts`:
The `scripts` directory will soon no longer exist, because the scripts and
their command line interfaces (CLIs) are being integrated into the python
package directly, which is the preferred method for deploying command line
accessible scripts -- it also has the nice side effect of making absolutely all
of their potential functionality available in code (which is nice for e.g.
updating the datastore in parallel for several years or data sources).

### The Secret Place: `pudl.package_data`

There are some files from the repository that are not code, but which we need
to distribute with the library for the code to run, and for users to use.

The ones that need to be accessible to and editable by the user get copied out
of The Secret Place by the `setup()` function mentioned above. The ones that
should not be edited by or accessible to the user on an installed system remain
in a special directory that is inside the pudl python package. All of these
files / information that are distributed with the python package but which are
not code are called “package data” and so I have called the directory
“package_data” which is of course frustratingly similar to datapackage. But
oh well.

Because python packages can be installed in a variety of ways -- sometimes as
distinct files the way we store them in the repository, and sometimes as
zipfiles or pre-compiled binaries -- the files that are stored in package_data
may not really be “files” at all, and a more abstract way of accessing them is
required.  As of Python 3.7 the official way of doing this is with the
[`importlib.resources`](https://docs.python.org/3/library/importlib.html#module-importlib.resources) module.

As a result, we need to treat access to the resources in the Secret Place
differently from resources that are deployed on the user’s system.  In cases
where we need to treat them similarly or the same, we can use the
`importlib.resources.path()` function to return a context manager that yields a
pathlib.Path object which can then be opened in the same way we would open
other files.

Because of this it’s important going forward that we clearly differentiate
between functions that are trying to look at a file in user space, and those
that are getting a resource from the Secret Place.

I renamed the all of the `input_validate` functions to `validate_input` and
made the top level one that's not associated with any data source public
(removed the leading `_underscore`).

I added a couple of new modules (`pudl.etl_pkg` and `pudl.glue.ferc1_eia`) to
the top level `__init__.py` to make sure they are available when someone does
`import pudl`

I renamed all the ETL_ functions and modules to use only `lower_case`.  Partly
this is a Python language convention, and partly it’s to avoid nasty filesystem
conflicts, where some filesystems allow case differences to be treated as
different files, and others don’t (e.g etl_pkg.py and ETL_pkg.py would be the
*same* file in some cases, and *different* files in other cases, which can make
for icky cross-platform behavior where necessary files accidentally get
clobbered.

Because the various `_etl_datasource` functions require different inputs, I
broke them into different function calls… e.g. FERC requires information both
about where the datastore is and where the sqlite outputs are, EIA requires
only the location of the datastore, and the glue ETL doesn’t need anything,
since it gets its inputs from the Secret Place. Giving all of them the full
`pudl_settings` as an input seemed deceptive, especially since it seemed like
it was just a way to enable a uniform looping construct over the data sources
in the top level `etl_pkg` function.

However, I could be convinced otherwise on this, especially if we can see more
datasets generally needing access to more information from the `pudl_settings`
dictionary for some reasons.  I could also be convinced that FERC should only
get a couple of inputs picked from `pudl_settings`, rather than the whole
dictionary as it is right now. I think it only needs 2 values, total, and
generally, it seems like pruning the values that are passed onward as early as
possible is a good idea for being able to track dependencies.

+++ src/pudl/output/export.py
@@ -8,7 +8,9 @@ import datetime
 import hashlib
 import logging
 import os
+import pathlib
 import re
+import uuid
 
 import datapackage
 import goodtables
@@ -17,6 +19,7 @@ import sqlalchemy as sa
 import tableschema
 
 import pudl
+from pudl import constants as pc
 
 logger = logging.getLogger(__name__)
 
@@ -94,7 +97,6 @@ def get_fields(table):
             newfield['constraints'] = {'enum': table.c[col].type.enums}
         if table.c[col].comment:
             newfield['description'] = table.c[col].comment
-
         fields.append(newfield)
     return fields
 
@@ -223,7 +225,7 @@ def get_tabular_data_resource(tablename, pkg_dir, testing=False):
     for src in data_sources:
         if src in pudl.constants.data_sources:
             descriptor["sources"].append({"title": src,
-                                          "path": "idfk"})
+                                          "path": pc.base_data_urls[src]})
 
     resource = datapackage.Resource(descriptor)
     if not resource.valid:
@@ -245,18 +247,23 @@ def hash_csv(csv_path):
 
     Returns the hexdigest of the hash, with a sha1: prefix as a string.
     """
+    # how big of a bit should I take?
     blocksize = 65536
+    # sha1 is a function of hashlib, not super accurate but its fast
     hasher = hashlib.sha1()
+    # opening the file and eat it for lunch
     with open(csv_path, 'rb') as afile:
         buf = afile.read(blocksize)
         while len(buf) > 0:
             hasher.update(buf)
             buf = afile.read(blocksize)
 
+    # returns the hash
     return f"sha1:{hasher.hexdigest()}"
 
 
-def data_package(pkg_tables, pkg_skeleton, pudl_settings=None, testing=False):
+def data_package(pkg_tables, pkg_skeleton, pudl_settings=None,
+                 testing=False, dry_run=False):
     """
     Create a data package of requested tables and their dependencies.
 
@@ -307,11 +314,16 @@ def data_package(pkg_tables, pkg_skeleton, pudl_settings=None, testing=False):
     for t in all_tables:
         csv_out = os.path.join(data_dir, f"{t}.csv")
         os.makedirs(os.path.dirname(csv_out), exist_ok=True)
-        df = pd.read_sql_table(t, pudl.init.connect_db(testing=testing))
-        if t in pudl.constants.need_fix_inting:
-            df = pudl.helpers.fix_int_na(df, pudl.constants.need_fix_inting[t])
-        logger.info(f"Exporting {t} to {csv_out}")
-        df.to_csv(csv_out, index=False)
+        if dry_run is True:
+            logger.info(f"Skipping export of {t} to {csv_out}")
+            pathlib.Path(csv_out).touch()
+        else:
+            df = pd.read_sql_table(t, pudl.init.connect_db(testing=testing))
+            if t in pudl.constants.need_fix_inting:
+                df = pudl.helpers.fix_int_na(
+                    df, pudl.constants.need_fix_inting[t])
+                logger.info(f"Exporting {t} to {csv_out}")
+                df.to_csv(csv_out, index=False)
 
     # Create a tabular data resource for each of the tables.
     resources = []
@@ -358,10 +370,11 @@ def data_package(pkg_tables, pkg_skeleton, pudl_settings=None, testing=False):
 
     data_pkg.save(pkg_json)
 
-    # Validate the data within the package using goodtables:
-    report = goodtables.validate(pkg_json, row_limit=100_000)
-    if not report['valid']:
-        logger.warning("Data package data validation failed.")
+    if not dry_run:
+        # Validate the data within the package using goodtables:
+        report = goodtables.validate(pkg_json, row_limit=100_000)
+        if not report['valid']:
+            logger.warning("Data package data validation failed.")
 
     return data_pkg
 
@@ -431,3 +444,189 @@ def annotated_xlsx(df, notes_dict, tags_dict, first_cols, sheet_name,
     # Return the xlsx_writer object, which can be written out, outside of
     # function, with 'xlsx_writer.save()'
     return xlsx_writer
+
+###############################################################################
+# CREATING PACKAGES AND METADATA
+###############################################################################
+
+
+def test_file_consistency(pkg_name, tables, out_dir):
+    """
+    Testing the consistency of tables for packaging
+
+    The purpose of this function is to test that we have the correct list of
+    tables.There are three different ways we could determine which tables are
+    being dumped into packages: a list of the tabels being generated through
+    the ETL functions, the list of dependent tables and the list of CSVs in
+    package directory.
+    """
+    file_tbls = [x.replace(".csv", "") for x in os.listdir(
+        os.path.join(out_dir, pkg_name, 'data'))]
+    dependent_tbls = list(
+        pudl.helpers.get_dependent_tables_from_list_pkg(tables)
+    )
+    etl_tbls = tables
+
+    dependent_tbls.sort()
+    file_tbls.sort()
+    etl_tbls.sort()
+    # TODO: determine what to do about the dependent_tbls... right now the
+    # dependent tables include some glue tables for FERC in particular, but
+    # we are imagining the glue tables will be
+    if ((file_tbls == etl_tbls)):  # & (dependent_tbls == etl_tbls)):
+        logger.info(f"Tables are consistent for {pkg_name} package")
+    else:
+        inconsistent_tbls = []
+        for tbl in file_tbls:
+            if tbl not in etl_tbls:
+                inconsistent_tbls.extend(tbl)
+                raise AssertionError(f"{tbl} from CSVs not in ETL tables")
+
+        # for tbl in dependent_tbls:
+        #    if tbl not in etl_tbls:
+        #        inconsistent_tbls.extend(tbl)
+        #        raise AssertionError(
+        #            f"{tbl} from forgien key relationships not in ETL tables")
+        # this is here for now just in case the previous two asserts don't work..
+        # we should probably just stick to one.
+        raise AssertionError(
+            f"Tables are inconsistent. "
+            f"Missing tables include: {inconsistent_tbls}")
+
+
+def get_tabular_data_resource_2(table_name, pkg_dir, testing=False):
+    """
+    Create a Tabular Data Resource descriptor for a PUDL table.
+
+    Based on the information in the database, and some additional metadata,
+    stored elsewhere (Where?!?!) this function will generate a valid Tabular
+    Data Resource descriptor, according to the Frictionless Data specification,
+    which can be found here:
+
+    https://frictionlessdata.io/specs/tabular-data-resource/
+    """
+    # Where the CSV file holding the data is, relative to datapackage.json
+    # This is the value that has to be embedded in the data package.
+    csv_relpath = os.path.join('data', f'{table_name}.csv')
+    # We need to access the file to calculate hash and size too:
+    csv_abspath = os.path.join(os.path.abspath(pkg_dir), csv_relpath)
+
+    # pull the skeleton of the descriptor from the megadata file
+    descriptor = pudl.helpers.pull_resource_from_megadata(table_name)
+    descriptor['path'] = csv_relpath
+    descriptor['bytes'] = os.path.getsize(csv_abspath)
+    descriptor['hash'] = pudl.output.export.hash_csv(csv_abspath)
+    descriptor['created'] = (datetime.datetime.utcnow().
+                             replace(microsecond=0).isoformat() + 'Z'),
+
+    resource = datapackage.Resource(descriptor)
+    if resource.valid:
+        logger.debug(f"{table_name} is a valid resource")
+    if not resource.valid:
+        raise AssertionError(
+            f"""
+            Invalid tabular data resource: {resource.name}
+
+            Errors:
+            {resource.errors}
+            """
+        )
+
+    return descriptor
+
+
+def generate_metadata(pkg_settings, tables, pkg_dir,
+                      uuid_pkgs=uuid.uuid4()):
+    # pkg_json is the datapackage.json that we ultimately output:
+    pkg_json = os.path.join(pkg_dir, "datapackage.json")
+    # Create a tabular data resource for each of the tables.
+    resources = []
+    for t in tables:
+        resources.append(
+            get_tabular_data_resource_2(t, pkg_dir=pkg_dir))
+
+    data_sources = pudl.helpers.data_sources_from_tables_pkg(
+        tables)
+    sources = []
+    for src in data_sources:
+        if src in pudl.constants.data_sources:
+            sources.append({"title": src,
+                            "path": pc.base_data_urls[src]})
+
+    contributors = set()
+    for src in data_sources:
+        for c in pudl.constants.contributors_by_source[src]:
+            contributors.add(c)
+
+    pkg_descriptor = {
+        "name": pkg_settings["name"],
+        "profile": "tabular-data-package",
+        "title": pkg_settings["title"],
+        "id": uuid_pkgs,
+        "description": pkg_settings["description"],
+        # "keywords": pkg_settings["keywords"],
+        "homepage": "https://catalyst.coop/pudl/",
+        "created": (datetime.datetime.utcnow().
+                    replace(microsecond=0).isoformat() + 'Z'),
+        "contributors": [pudl.constants.contributors[c] for c in contributors],
+        "sources": sources,
+        "licenses": [pudl.constants.licenses["cc-by-4.0"]],
+        "resources": resources,
+    }
+
+    # Use that descriptor to instantiate a Package object
+    data_pkg = datapackage.Package(pkg_descriptor)
+
+    # Validate the data package descriptor before we go to
+    if not data_pkg.valid:
+        logger.warning(f"""
+            Invalid tabular data package: {data_pkg.descriptor["name"]}
+            Errors: {data_pkg.errors}""")
+
+    data_pkg.save(pkg_json)
+    # Validate the data within the package using goodtables:
+    report = goodtables.validate(pkg_json, row_limit=1000)
+    if not report['valid']:
+        logger.warning("Data package data validation failed.")
+
+    return data_pkg, report
+
+
+def generate_data_packages(settings, pudl_settings, debug=False):
+    """
+    Cordinate the generation of data packages.
+
+    Args:
+        settings (iterable) : a list of dictionaries. Each item in the list
+            corresponds to a data package. Each data package's dictionary
+            contains the arguements for its ETL function.
+        out_dir (path-like): The location of the packaging directory.
+            The data package will be created in a subdirectory in
+            this directory, according to the name of the package.
+        debug (boolean)
+    """
+    # validate the settings from the settings file.
+    validated_settings = pudl.etl_pkg.validate_input(settings)
+    uuid_pkgs = str(uuid.uuid4())
+    metas = {}
+    for pkg in validated_settings:
+        # run the ETL functions for this pkg and return the list of tables
+        # dumped to CSV
+        pkg_tbls = pudl.etl_pkg.etl_pkg(pkg, pudl_settings)
+        # assure that the list of tables from ETL match up with the CVSs and
+        # dependent tables
+        test_file_consistency(pkg['name'],
+                              pkg_tbls,
+                              out_dir=pudl_settings['datapackage_dir'])
+        # generate the metadata for the package and validate
+        # TODO: we'll probably want to remove this double return... but having
+        # the report and the metadata while debugging is very useful.
+        meta, report = generate_metadata(
+            pkg,
+            pkg_tbls,
+            os.path.join(pudl_settings['datapackage_dir'], pkg['name']),
+            uuid_pkgs=uuid_pkgs
+        )
+        metas[pkg['name']] = [meta, report]
+    if debug:
+        return (metas)
